MainActivity를 WorldMap으로 바꿈.(작업하던 것들 다 합쳐서 필요없는 파일 지움) WorldMap완성
