+++ .idea/gradle.xml
@@ -5,16 +5,11 @@
       <GradleProjectSettings>
         <option name="distributionType" value="DEFAULT_WRAPPED" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
-        <option name="gradleHome" value="$APPLICATION_HOME_DIR$/gradle/gradle-2.10" />
         <option name="gradleJvm" value="1.7" />
         <option name="modules">
           <set>
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
-<<<<<<< HEAD
-            <option value="$PROJECT_DIR$/parallaxviewpager" />
-=======
->>>>>>> 7b246b2... [추가] 여형경보 지도 가져오기 추가
           </set>
         </option>
       </GradleProjectSettings>
