+++ app/src/main/AndroidManifest.xml
@@ -21,6 +21,7 @@
         <activity android:name=".WorldMap"/>
         <activity android:name=".Country_info"/>
         <activity android:name=".fragment.ListActivity"/>
+
         <meta-data
             android:name="com.google.android.geo.API_KEY"
             android:value="AIzaSyAsxByWSpI8bwTR6xP2Lq9G5LdiKBkTWR8"/>
