--- app/src/main/java/com/example/cheonyujung/accidentofworld/fragment/mapActivity.java
@@ -1,20 +1,0 @@
-//package com.example.cheonyujung.accidentofworld.fragment;
-//
-//import android.app.FragmentManager;
-//import android.os.Bundle;
-//
-//import com.example.cheonyujung.accidentofworld.R;
-//
-///**
-// * Created by vmffkxlgnqh1 on 2016. 5. 28..
-// */
-//public class mapActivity extends singleFragmentActivity{
-//    @Override
-//    protected void onCreate(Bundle savedInstanceState) {
-//        super.onCreate(savedInstanceState);
-//        FragmentManager fm = getFragmentManager();
-//        fm.beginTransaction()
-//                .replace(R.id.body, com.example.cheonyujung.accidentofworld.fragment.WorldMapFragment.getInstence())
-//                .commit();
-//    }
-//}
